Model Identification/requirements.txt
streamlit
numpy
pillow
joblib
scikit-learn
tensorflow